requirements.txt
kafka-python==1.3.3
tweepy==3.5.0
